UIManager fix 3

+++ Contract, UI & Routing/UIManager/Handlers/RegisterUIFragmentCommandHandler.cs
@@ -34,7 +34,7 @@ namespace SoftwareCenter.UIManager.Handlers
                 command.HtmlContent,
                 command.CssContent,
                 command.JsContent,
-                command.Priority,
+                (int)command.Priority,
                 command.SlotName
             );
